Instantiated ConfigMap and Secret labels
An EventBasedAddOn can reference ConfigMaps and Secrets.
When an event happens, the EventBasedAddOn controller creates one or more
ClusterProfile (depending on its configuration).
Those referenced ConfigMaps and Secrets can be expressed as template and
instantiated using values from resources in managed clusters that generated
the event.

Each ConfigMap, Secret referenced by EventBasedAddOn gets instantiated in
the projectsveltos namespace. And the created ClusterProfile references that.

When creating a ConfigMap or Secret, labels are added. Before this PR labels
only contained information on the managed cluster and the EventBasedAddOn instance
name.

This PR adds two more labels containing the original ConfigMap or Secret namespace
and name.
The labels are also used to generated the name of the instantiated ConfigMap or Secret.

This fixes #94. Such issues consisted that only one ConfigMap or Secret was created
even if the EventBasedAddOn was referencing more than one ConfigMap or more than one
Secret. Issue was tied to generated name which did not contain any info regarding the
original referenced object.