(feat) CloudEvents
EventSource can match CloudEvents over NATS/JetStream subjects.

EventReports in the managed cluster contains the matching CloudEvents.
When EventReport instances are collected, they are processed in the management
cluster. If all is successfull, EventReport is stored in the management
cluster (with no matching cloudEvents, those are removed before storing it).
The EventReport is updated in the managed cluster by removing the processed
CloudEvents and setting the status to Processed.

If an error happens, EventReport in the managed cluster is not updated.
So it will be fetched again and again the matching CloudEvents will be
processed.

ClusterProfiles/Secrets/ConfigMaps created because of CloudEvents have two labels:
1. eventtrigger.lib.projectsveltos.io/cesource containing the CloudEvent Source
1. eventtrigger.lib.projectsveltos.io/cesubject containing the CloudEvent Subject

Resources created because of a CloudEvent are removed only in those scenario:
1. EventTrigger is removed
2. Cluster stops being a match for an EventTrigger
3. EventTrigger.Spec.CloudEventAction is set to ``Delete`

The last case is important. It means Sveltos identifies CloudEvents using
Source,Subject pair.