(bug) duplicate ClusterProfiles
In v0.52.2, when running in agentless mode, the EventReport name is different.
It used to be simply the EventSource name, but now it is different.
EventReport name is one of the labels added to CLusterProfiles created
by event-manager in response to event.s

This discrepancy led to duplicate ClusterProfile instantiations after Sveltos
upgrades to v0.52.2

To resolve this, we leverage the projectsveltos.io/eventsource-name label
present on EventReport objects for identification, instead of relying on
the EventReport's name directly.

The existing 'eventReportNameLabel' is now technically inaccurate as it
stores the EventSource name, not the EventReport name. To clarify this,
we are introducing 'eventSourceNameLabel'.

We plan to remove 'eventReportNameLabel' in a few releases following v0.52.2.