(feat) Remove Json marshaling
When instantiating referenced resource, Sveltos fecthes the content
of referenced ConfigMap/Secret instances. The result is a map[string]string

The only section that is a template, is the value (keys can be assumed to
be constant). This PR replaces the json.Marshal of the string with code
that walks of keys in the map, and instantiated the values.